cleanup, more actor blocking debugging
